+++ Core/GDCore/Project/ResourcesLoader.cpp
@@ -1,7 +1,7 @@
 #include "GDCore/Project/ResourcesLoader.h"
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
-#include "GDCore/Tools/FStream.h"
+#include "GDCore/Tools/FileStream.h"
 #include "GDCore/String.h"
 #include <fstream>
 #include <iostream>
@@ -18,8 +18,6 @@ ResourcesLoader * ResourcesLoader::_singleton = NULL;
 
 void ResourcesLoader::LoadSFMLImage( const gd::String & filename, sf::Image & image )
 {
-    std::wcout << L"Trying to load" << filename.ToWide() << std::endl;
-
     gd::SFMLFileStream stream;
     if (!stream.open(filename) || !image.loadFromStream(stream))
         cout << "Failed to load a SFML image: " << filename << endl;
@@ -35,8 +33,6 @@ sf::Texture ResourcesLoader::LoadSFMLTexture(const gd::String & filename)
 
 void ResourcesLoader::LoadSFMLTexture( const gd::String & filename, sf::Texture & texture )
 {
-    std::wcout << L"Trying to load" << filename.ToWide() << std::endl;
-
     gd::SFMLFileStream stream;
     if (!stream.open(filename) || !texture.loadFromStream(stream))
         cout << "Failed to load a SFML texture: " << filename << endl;
@@ -91,9 +87,9 @@ gd::String ResourcesLoader::LoadPlainText( const gd::String & filename )
  */
 char* ResourcesLoader::LoadBinaryFile( const gd::String & filename )
 {
-    ifstream file (filename.ToLocale().c_str(), ios::in|ios::binary|ios::ate);
+    gd::FileStream file (filename, ios::in|ios::binary|ios::ate);
     if (file.is_open()) {
-        ifstream::pos_type size = file.tellg();
+        iostream::pos_type size = file.tellg();
         char * memblock = new char [size];
         file.seekg (0, ios::beg);
         file.read (memblock, size);
@@ -108,7 +104,7 @@ char* ResourcesLoader::LoadBinaryFile( const gd::String & filename )
 
 long int ResourcesLoader::GetBinaryFileSize( const gd::String & filename)
 {
-    ifstream file (filename.ToLocale().c_str(), ios::in|ios::binary|ios::ate);
+    gd::FileStream file (filename, ios::in|ios::binary|ios::ate);
     if (file.is_open()) {
         return file.tellg();
     }
