+++ Core/GDCore/Tools/FileStream.h
